feat: Optimize database schema and enable compression
This commit optimizes the database schema for space efficiency and enables PostgreSQL's LZ4 compression.

- Tightened data types across all schema files, using more specific types like `REAL`, `DATE`, and `VARCHAR(n)` instead of `DOUBLE PRECISION` and `TEXT`.
- Enabled `lz4` compression in the `docker-compose.yml` for the PostgreSQL service to reduce storage for TOAST-able data.
- Addressed data integrity issues by reverting specific columns to `BIGINT` and `DOUBLE PRECISION` where necessary to prevent data overflow and precision loss.

+++ db/20-create-sde-map-tables.sql
@@ -9,9 +9,9 @@ CREATE TABLE IF NOT EXISTS regions (
     "description" TEXT,
     "factionId" INTEGER,
     "nebulaId" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "wormholeClassId" INTEGER,
     "updatedAt" TIMESTAMP DEFAULT NOW(),
     "version" INTEGER
@@ -23,9 +23,9 @@ CREATE TABLE IF NOT EXISTS constellations (
     "name" VARCHAR(255),
     "regionId" INTEGER,
     "factionId" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "solarSystemIds" INTEGER[],
     "wormholeClassId" INTEGER,
     "updatedAt" TIMESTAMP DEFAULT NOW(),
@@ -46,9 +46,9 @@ CREATE TABLE IF NOT EXISTS solarSystems (
     "international" BOOLEAN,
     "luminosity" REAL,
     "planetIds" INTEGER[],
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "radius" DOUBLE PRECISION,
     "regional" BOOLEAN,
     "securityClass" VARCHAR(255),
@@ -68,9 +68,9 @@ CREATE TABLE IF NOT EXISTS stargates (
     "solarSystemId" INTEGER,
     "destinationGateId" BIGINT,
     "destinationSolarSystemId" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "typeId" INTEGER,
     "version" INTEGER
 );
@@ -96,9 +96,9 @@ CREATE TABLE IF NOT EXISTS planets (
     "solarSystemId" INTEGER,
     "typeId" INTEGER,
     "celestialIndex" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "version" INTEGER
 );
 
@@ -110,9 +110,9 @@ CREATE TABLE IF NOT EXISTS moons (
     "planetId" INTEGER,
     "typeId" INTEGER,
     "celestialIndex" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "version" INTEGER
 );
 
@@ -123,8 +123,8 @@ CREATE TABLE IF NOT EXISTS asteroidBelts (
     "solarSystemId" INTEGER,
     "typeId" INTEGER,
     "celestialIndex" INTEGER,
-    "positionX" REAL,
-    "positionY" REAL,
-    "positionZ" REAL,
+    "positionX" DOUBLE PRECISION,
+    "positionY" DOUBLE PRECISION,
+    "positionZ" DOUBLE PRECISION,
     "version" INTEGER
 );
